OP-1216 Use BigDecimal for the transient MedicalWard quantity
Change MedicalWard.qty (a @Transient computed aggregate) from Double to BigDecimal
so the ward-pharmacy quantity path stays exact end-to-end instead of round-tripping
through double. The ward-stock computation now subtracts BigDecimals directly.
Other consumers that keep their own double-based model (MedicalInventoryManager,
the Jasper print DTO MedicalWardForPrint) convert at the boundary, so no report or
inventory behaviour changes.

+++ src/main/java/org/isf/medicalinventory/manager/MedicalInventoryManager.java
@@ -471,7 +471,7 @@ public class MedicalInventoryManager {
 			Optional<MedicalWard> optMedicalWard = movWardBrowserManager.getMedicalsWard(inventory.getWardCode(), medical.getCode(), false).stream()
 				.filter(m -> m.getLot().getCode().equals(lotCode)).findFirst();
 
-			double wardStoreQty = optMedicalWard.isPresent() ? optMedicalWard.get().getQty() : 0.0;
+			double wardStoreQty = optMedicalWard.isPresent() ? optMedicalWard.get().getQty().doubleValue() : 0.0;
 
 			// Search for the specific Lot and Medical in inventoryRowSearchList
 			Optional<MedicalInventoryRow> matchingRow = inventoryRowSearchList.stream()
@@ -881,7 +881,7 @@ public class MedicalInventoryManager {
 			// Fetch also empty lots because some movements may have discharged them completely
 			Optional<MedicalWard> optMedicalWard = movWardBrowserManager.getMedicalsWard(inventory.getWardCode(), medical.getCode(), false).stream()
 				.filter(m -> m.getLot().getCode().equals(lotCode)).findFirst();
-			double wardStoreQty = optMedicalWard.isPresent() ? optMedicalWard.get().getQty() : 0.0;
+			double wardStoreQty = optMedicalWard.isPresent() ? optMedicalWard.get().getQty().doubleValue() : 0.0;
 
 			// Search for the specific Lot and Medical in inventoryRowSearchList (Lot should be enough)
 			Optional<MedicalInventoryRow> matchingRow = inventoryRowList.stream()
